+++ ZipUtilities/NOZUnzipper.m
@@ -111,18 +111,22 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
 
 - (BOOL)openAndReturnError:(out NSError **)error
 {
-    NSError *stackError = NOZError(NOZErrorCodeUnzipCannotOpenZip, @{ @"zipFilePath" : [self zipFilePath] ?: [NSNull null] });
+    NSError *stackError = NOZErrorCreate(NOZErrorCodeUnzipCannotOpenZip, @{ @"zipFilePath" : [self zipFilePath] ?: [NSNull null] });
     _standardizedFilePath = [self.zipFilePath stringByStandardizingPath];
     if (_standardizedFilePath.UTF8String) {
         _internal.file = fopen(_standardizedFilePath.UTF8String, "r");
         if (_internal.file) {
-            _internal.endOfFilePosition = fseeko(_internal.file, 0, SEEK_END);
+            if (0 == fseeko(_internal.file, 0, SEEK_END)) {
+                _internal.endOfFilePosition = ftello(_internal.file);
+            } else {
+                _internal.endOfFilePosition = (off_t)[[[NSFileManager defaultManager] attributesOfItemAtPath:_standardizedFilePath error:nil] fileSize];
+            }
             _internal.endOfCentralDirectorySignaturePosition = [self private_locateSignature:NOZMagicNumberEndOfCentralDirectoryRecord];
             if (_internal.endOfCentralDirectorySignaturePosition) {
                 return YES;
             } else {
                 [self closeAndReturnError:NULL];
-                stackError = NOZError(NOZErrorCodeUnzipInvalidZipFile, @{ @"zipFilePath" : self.zipFilePath });
+                stackError = NOZErrorCreate(NOZErrorCodeUnzipInvalidZipFile, @{ @"zipFilePath" : self.zipFilePath });
             }
         }
     }
@@ -153,7 +157,7 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
     });
 
     if (!_internal.file || !_internal.endOfCentralDirectorySignaturePosition) {
-        stackError = NOZError(NOZErrorCodeUnzipMustOpenUnzipperBeforeManipulating, nil);
+        stackError = NOZErrorCreate(NOZErrorCodeUnzipMustOpenUnzipperBeforeManipulating, nil);
         return nil;
     }
 
@@ -161,12 +165,12 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
 
     @autoreleasepool {
         if (![cd readEndOfCentralDirectoryRecordAtPosition:_internal.endOfCentralDirectorySignaturePosition inFile:_internal.file]) {
-            stackError = NOZError(NOZErrorCodeUnzipCannotReadCentralDirectory, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipCannotReadCentralDirectory, nil);
             return nil;
         }
 
         if (![cd readCentralDirectoryEntriesWithFile:_internal.file]) {
-            stackError = NOZError(NOZErrorCodeUnzipCannotReadCentralDirectory, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipCannotReadCentralDirectory, nil);
             return nil;
         }
 
@@ -183,7 +187,7 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
 {
     if (index >= _centralDirectory.recordCount) {
         if (error) {
-            *error = NOZError(NOZErrorCodeUnzipIndexOutOfBounds, nil);
+            *error = NOZErrorCreate(NOZErrorCodeUnzipIndexOutOfBounds, nil);
         }
         return nil;
     }
@@ -214,24 +218,24 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
 
     @autoreleasepool {
         if (!_internal.file) {
-            stackError = NOZError(NOZErrorCodeUnzipMustOpenUnzipperBeforeManipulating, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipMustOpenUnzipperBeforeManipulating, nil);
             return NO;
         }
 
         if (![record isOwnedByCentralDirectory:_centralDirectory]) {
-            stackError = NOZError(NOZErrorCodeUnzipCannotReadFileEntry, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipCannotReadFileEntry, nil);
             return NO;
         }
 
         if (![self private_locateCompressedDataOfRecord:record]) {
-            stackError = NOZError(NOZErrorCodeUnzipCannotReadFileEntry, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipCannotReadFileEntry, nil);
             return NO;
         }
 
         do {
             NOZErrorCode code = [record validate];
             if (0 != code) {
-                stackError = NOZError(code, nil);
+                stackError = NOZErrorCreate(code, nil);
                 return NO;
             }
         } while (0);
@@ -241,7 +245,7 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
 
         _currentUnzipping.offsetToFirstByte = ftello(_internal.file);
         if (_currentUnzipping.offsetToFirstByte == -1) {
-            stackError = NOZError(NOZErrorCodeUnzipCannotReadFileEntry, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipCannotReadFileEntry, nil);
             return NO;
         }
 
@@ -264,12 +268,12 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
         });
 
         if (!_currentDecoder || !_currentDecoderContext) {
-            stackError = NOZError(NOZErrorCodeUnzipDecompressionMethodNotSupported, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipDecompressionMethodNotSupported, nil);
             return NO;
         }
         
         if (![_currentDecoder initializeDecoderContext:_currentDecoderContext]) {
-            stackError = NOZError(NOZErrorCodeUnzipFailedToDecompressEntry, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipFailedToDecompressEntry, nil);
             return NO;
         }
         
@@ -281,7 +285,7 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
         }
         
         if (![_currentDecoder finalizeDecoderContext:_currentDecoderContext]) {
-            stackError = NOZError(NOZErrorCodeUnzipFailedToDecompressEntry, nil);
+            stackError = NOZErrorCreate(NOZErrorCodeUnzipFailedToDecompressEntry, nil);
             return NO;
         }
     }
@@ -545,7 +549,7 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
     __block BOOL success = YES;
     noz_defer(^{
         if (!success && error && !*error) {
-            *error = NOZError(NOZErrorCodeUnzipCannotDecompressFileEntry, nil);
+            *error = NOZErrorCreate(NOZErrorCodeUnzipCannotDecompressFileEntry, nil);
         }
     });
 
@@ -592,7 +596,7 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
     if (_currentUnzipping.crc32 != _currentUnzipping.entry->fileDescriptor.crc32) {
         success = NO;
         if (error) {
-            *error = NOZError(NOZErrorCodeUnzipChecksumMissmatch, nil);
+            *error = NOZErrorCreate(NOZErrorCodeUnzipChecksumMissmatch, nil);
         }
 
         return NO;
@@ -625,7 +629,7 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
 - (instancetype)initWithKnownFileSize:(SInt64)fileSize
 {
     if (self = [super init]) {
-        _totalUncompressedSize = fileSize;
+        _totalCompressedSize = fileSize;
     }
     return self;
 }
@@ -786,7 +790,7 @@ static BOOL noz_fread_value(FILE *file, Byte* value, const UInt8 byteCount);
     __block NSDictionary *userInfo = nil;
     noz_defer(^{
         if (code && error) {
-            *error = NOZError(code, userInfo);
+            *error = NOZErrorCreate(code, userInfo);
         }
     });